SSOR-50 Added password recovery

+++ boss-user/src/main/java/org/ssor/boss/user/controller/UserController.java
@@ -118,34 +118,23 @@ package org.ssor.boss.user.controller;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.annotation.Secured;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.util.MimeTypeUtils;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+import org.ssor.boss.core.service.AwsSesService;
 import org.ssor.boss.core.service.ConfirmationService;
 import org.ssor.boss.core.service.UserService;
-import org.ssor.boss.core.transfer.ApiRequestResponse;
-import org.ssor.boss.core.transfer.FinalizeConfirmationInput;
-import org.ssor.boss.core.transfer.RegisterUserInput;
-import org.ssor.boss.core.transfer.RegisterUserOutput;
-import org.ssor.boss.core.transfer.SecureUserDetails;
-import org.ssor.boss.core.transfer.UpdateUserInput;
+import org.ssor.boss.core.transfer.*;
 import org.ssor.boss.user.dto.ForgotPassEmailInput;
 import org.ssor.boss.user.dto.ForgotPassTokenInput;
 import org.ssor.boss.user.service.ControllerService;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Optional;
+
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.OK;
 import static org.springframework.util.MimeTypeUtils.APPLICATION_JSON_VALUE;
@@ -155,14 +144,15 @@ import static org.ssor.boss.core.entity.ConfirmationType.USER_CONFIRMATION;
 @Validated
 @RestController
 @AllArgsConstructor
+@CrossOrigin
 @RequestMapping(value = UserController.USERS_ROUTE)
 public class UserController
 {
   public static final String USERS_ROUTE = "/api/v1/users";
   private final UserService userService;
+  private final AwsSesService awsSesService;
   private final ConfirmationService confirmationService;
   private final ControllerService controllerService;
-  // private final MailService mailService;
 
   @GetMapping
   public ResponseEntity<List<SecureUserDetails>> getAllUsers()
@@ -174,9 +164,9 @@ public class UserController
   public ResponseEntity<RegisterUserOutput> registerNewUser(@Valid @RequestBody RegisterUserInput registerUserInput)
   {
     final var newUser = userService.registerNewUser(registerUserInput, Instant.now());
-    final var generationOutput = confirmationService.generateConfirmation(USER_CONFIRMATION, newUser.getId());
+    //final var generationOutput = confirmationService.generateConfirmation(USER_CONFIRMATION, newUser.getId());
     // Send email using confirmation generation output.
-    return ResponseEntity.ok(newUser);
+    return ResponseEntity.status(HttpStatus.CREATED).body(newUser);
   }
 
   @PostMapping("/confirmation")
@@ -201,7 +191,7 @@ public class UserController
   @GetMapping("/{user_id}")
   public ResponseEntity<Object> getUserById(@PathVariable("user_id") int userId, HttpServletRequest request)
   {
-    final var optionalUser = userService.getSecureUserDetailsWithId(userId);
+    var optionalUser = userService.getSecureUserDetailsWithId(userId);
     if (optionalUser.isPresent())
       return ResponseEntity.ok(optionalUser.get());
 
@@ -215,7 +205,8 @@ public class UserController
     return ResponseEntity.status(NOT_FOUND).body(response);
   }
 
-  @PutMapping("/{user_id}")
+  @PutMapping(path ="/{user_id}",consumes = { APPLICATION_JSON_VALUE,
+          APPLICATION_XML_VALUE })
   public ResponseEntity<ApiRequestResponse>
   updateUserProfile(@PathVariable("user_id") int userId, @Valid @RequestBody UpdateUserInput updateUserInput,
                     HttpServletRequest request)
@@ -233,22 +224,30 @@ public class UserController
   public ResponseEntity<ApiRequestResponse>
   deleteUserWithId(@PathVariable("user_id") int userId, HttpServletRequest request)
   {
+    final var optionalUser = userService.getSecureUserDetailsWithId(userId);
+    if (optionalUser.isEmpty())
+      return notFoundResponse(request, String.format("User with id %d not found", userId));
+
     userService.deleteUserWithId(userId);
     return okResponse(request, String.format("User with id %d deleted", userId));
   }
 
-  @PostMapping(path = "/api/v1/user/email", consumes = { APPLICATION_JSON_VALUE,
+  @PostMapping(path = "/forgot-password", consumes = { APPLICATION_JSON_VALUE,
                                                          APPLICATION_XML_VALUE })
   public ResponseEntity<String> forgotPasswordEmail(
       @Valid @RequestBody ForgotPassEmailInput forgotPasswordEmailInput) {
-    controllerService.sendPasswordReset(forgotPasswordEmailInput);
+    Optional<Email> optEmail = controllerService.sendPasswordReset(forgotPasswordEmailInput);
+    if(optEmail.isEmpty()){
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("There is was an issue finding reset email.");
+    }
+    awsSesService.sendEmail(optEmail.get());
     return new ResponseEntity<>(null, HttpStatus.ACCEPTED);
   }
 
-  @PutMapping(path = "/api/v1/user/password", consumes = { APPLICATION_JSON_VALUE,
-                                                           APPLICATION_XML_VALUE })
+  @PutMapping(path = "/reset-password", consumes = { APPLICATION_JSON_VALUE,
+          APPLICATION_XML_VALUE })
   public ResponseEntity<String> updateForgotPassword(
-      @Valid @RequestBody ForgotPassTokenInput forgotPasswordTokenInput) {
+          @Valid @RequestBody ForgotPassTokenInput forgotPasswordTokenInput) {
     if (controllerService.updateForgotPassword(forgotPasswordTokenInput)) {
       return new ResponseEntity<>(null, HttpStatus.OK);
     }
@@ -266,7 +265,7 @@ public class UserController
     response.setMessage(message);
     response.setPath(request.getServletPath());
     response.setValidationErrors(new HashMap<>());
-    return ResponseEntity.ok(response);
+    return ResponseEntity.status(response.getStatus()).body(response);
   }
 
   private ResponseEntity<ApiRequestResponse> okResponse(HttpServletRequest request, String message)
